fixed a few compilation warnings
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@7859 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/dbase/fimonitor.cc
@@ -263,7 +263,7 @@ void HashCache::save()
 		std::cerr << "Cannot open encrypted file cache for writing: " << _path+".bin.tmp" << std::endl;
 		goto save_free;
 	}
-	if(fwrite(encryptedData,1,encDataLen,F) != encDataLen)
+    if(fwrite(encryptedData,1,encDataLen,F) != (uint32_t)encDataLen)
 	{
 		std::cerr << "Could not write entire encrypted hash cache file. Out of disc space??" << std::endl;
 		fclose(F) ;
